chore: bump version to 0.1.12; refactor imports and improve code organization
