+++ src/c_two/message/__init__.py
@@ -1,3 +1,3 @@
 from .client import Client
 from .server import Server
-from .transferable import register_transferable, get_transferable, serialize_from_table, deserialize_to_table, deserialize_to_rows
+from .util.arrow import serialize_from_table, deserialize_to_table, deserialize_to_rows
