+++ src/c_two/message/server.py
@@ -67,7 +67,7 @@ class Server:
             print(f'Error during termination: {e}')
             sys.exit(1)
             
-# Helper ##################################################
+# Helpers ##################################################
 
 def _parse_message(full_message: bytes) -> list[memoryview]:
     buffer = memoryview(full_message)
