+++ src/c_two/message/util/arrow.py
@@ -1,0 +1,28 @@
+import pyarrow as pa
+
+def serialize_from_table(table: pa.Table) -> bytes:
+    sink = pa.BufferOutputStream()
+    with pa.ipc.new_stream(sink, table.schema) as writer:
+        writer.write_table(table)
+    binary_data = sink.getvalue().to_pybytes()
+    return binary_data
+
+def deserialize_to_table(serialized_data: bytes) -> pa.Table:
+    buffer = pa.py_buffer(serialized_data)
+    with pa.ipc.open_stream(buffer) as reader:
+        table = reader.read_all()
+    return table
+
+def deserialize_to_rows(serialized_data: bytes) -> dict:
+    buffer = pa.py_buffer(serialized_data)
+
+    with pa.ipc.open_stream(buffer) as reader:
+        table = reader.read_all()
+
+    return table.to_pylist()
+
+arrow = {
+    'serialize_from_table': serialize_from_table,
+    'deserialize_to_table': deserialize_to_table,
+    'deserialize_to_rows': deserialize_to_rows
+}
